[chore #165538524] this commit renames the files, adds database for tests, removes some of the unwanted commented codes
