+++ .eslintrc.json
@@ -13,6 +13,17 @@
         "sourceType": "module"
     },
     "rules": {
-        "import/no-extraneous-dependencies": 0
+        "import/no-extraneous-dependencies": 0,
+        "consistent-return": 0,
+        "no-cond-assign": 0,
+        "no-multi-spaces": 0,
+        "no-restricted-globals": 0,
+        "radix": 0,
+        "template-curly-spacing": 0,
+        "no-unused-expressions": 0,
+        "import/prefer-default-export": 0,
+        "import/no-useless-path-segments": 0,
+        "no-shadow": 0,
+        "no-undef": 0
     }
 }
