+++ UI/HTML/admin.html
@@ -16,12 +16,6 @@ height: 100%;">
             <span class="open-slide">
                 <i class="fas fa-bars" onclick="showHiddenMenuArea()" id="barsMenu"></i>
             </span>
-            <!-- <ul class="navbar-nav">
-                <li><a href="#" onclick="showCreateAccount()">Create Account</a></li>
-                <li><a href="#" onclick="showTransaction()">Users' Accounts</a></li>
-                <li><a href="#" onclick="showResetPassword()">Reset Password</a></li>
-                <li><a href="../index.html" onclick="loginFunction()">Logout</a></li>
-            </ul> -->
        </div>
         <div class="navbar2">
             <span class="banka">
@@ -124,8 +118,7 @@ height: 100%;">
                         <td>GISA Chris</td>
                         <td>Saving</td>
                         <td>Active</td>
-                        <td>700,000 frw</td>
-                        <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
+                        <td>700,000 frw</t
                     </tr>
                     <tr onclick="viewAccountAdmin()">
                         <td>2</td>
@@ -134,8 +127,7 @@ height: 100%;">
                         <td>KAMI Christian</td>
                         <td>Saving</td>
                         <td>Active</td>
-                        <td>900,000 frw</td>
-                        <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
+                        <td>900,000 frw</t
                     </tr>
                     <tr onclick="viewAccountAdmin()">
                         <td>3</td>
@@ -144,8 +136,7 @@ height: 100%;">
                         <td>GAKWAYA Kevin</td>
                         <td>Current</td>
                         <td>Active</td>
-                        <td>1,000,000 frw</td>
-                        <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
+                        <td>1,000,000 frw</t
                     </tr>
                     <tr onclick="viewAccountAdmin()">
                         <td>4</td>
@@ -154,8 +145,7 @@ height: 100%;">
                         <td>GISA Christophe</td>
                         <td>Saving</td>
                         <td>Active</td>
-                        <td>700,000 frw</td>
-                        <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
+                        <td>700,000 frw</t
                     </tr>
                     <tr onclick="viewAccountAdmin()">
                         <td>5</td>
@@ -164,8 +154,7 @@ height: 100%;">
                         <td>GASANA Christophe</td>
                         <td>Current</td>
                         <td>Active</td>
-                        <td>700,000 frw</td>
-                        <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
+                        <td>700,000 frw</t
                     </tr>
                     <tr onclick="viewAccountAdmin()">
                         <td>6</td>
@@ -174,8 +163,7 @@ height: 100%;">
                         <td>GISANA Chris</td>
                         <td>Saving</td>
                         <td>Active</td>
-                        <td>800,000 frw</td>
-                        <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
+                        <td>800,000 frw</t
                     </tr>
 
                 </table>
@@ -250,7 +238,6 @@ height: 100%;">
                                  <td>Saving</td>
                                  <td>Active</td>
                                  <td>700,000 frw</td>
-                                 <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
                              </tr>
                              <tr onclick="viewAccountAdmin()">
                                  <td>2</td>
@@ -260,7 +247,6 @@ height: 100%;">
                                  <td>Saving</td>
                                  <td>Active</td>
                                  <td>900,000 frw</td>
-                                 <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
                              </tr>
                              <tr onclick="viewAccountAdmin()">
                                  <td>3</td>
@@ -270,7 +256,6 @@ height: 100%;">
                                  <td>Current</td>
                                  <td>Active</td>
                                  <td>1,000,000 frw</td>
-                                 <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
                              </tr>
                              <tr onclick="viewAccountAdmin()">
                                  <td>4</td>
@@ -280,7 +265,6 @@ height: 100%;">
                                  <td>Saving</td>
                                  <td>Active</td>
                                  <td>700,000 frw</td>
-                                 <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
                              </tr>
                              <tr onclick="viewAccountAdmin()">
                                  <td>5</td>
@@ -290,7 +274,6 @@ height: 100%;">
                                  <td>Current</td>
                                  <td>Active</td>
                                  <td>700,000 frw</td>
-                                 <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
                              </tr>
                              <tr onclick="viewAccountAdmin()">
                                  <td>6</td>
@@ -300,7 +283,6 @@ height: 100%;">
                                  <td>Saving</td>
                                  <td>Active</td>
                                  <td>800,000 frw</td>
-                                 <!-- <td><input type="submit" value="Activate"><input type="submit" value="Deactivate"><i class="fas fa-trash-alt"></i></td> -->
                              </tr>
          
                          </table>
