+++ dummy/helpers/signup.js
