+++ dummy/tests/sign.js
