+++ server/models/tables.js
@@ -1,4 +1,3 @@
-/* eslint-disable import/prefer-default-export */
 import pool from '../config/db';
