+++ server/routes/users.routes.js
@@ -1,6 +1,5 @@
 import express from 'express';
 import Users from '../controllers/users';
-// import auth from '../middleware/auth';
 
 const router = express.Router();
